Fix Windows build and small adjustments to Sys_MillisecondsPrecise()
for the nsec part of struct timespec a simple long is enough

+++ neo/sys/aros/aros_main.cpp
@@ -956,7 +956,7 @@ double Sys_MillisecondsPrecise() {
 	clock_gettime(CLOCK_MONOTONIC, &now);
 
 	long long sec = now.tv_sec - first.tv_sec;
-	long long nsec = now.tv_nsec - first.tv_nsec;
+	long nsec = now.tv_nsec - first.tv_nsec;
 
 	double ret = sec * 1000.0;
 	ret += double(nsec) * 0.000001;
